+++ app/src/main/java/com/example/taskfrog/ui/list/ListFragment.kt
@@ -1,4 +1,4 @@
-package com.example.taskfrog.ui.dashboard
+package com.example.taskfrog.ui.list
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -9,7 +9,7 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import com.example.taskfrog.databinding.FragmentDashboardBinding
 
-class DashboardFragment : Fragment() {
+class ListFragment : Fragment() {
 
     private var _binding: FragmentDashboardBinding? = null
 
@@ -23,7 +23,7 @@ class DashboardFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View {
         val dashboardViewModel =
-            ViewModelProvider(this).get(DashboardViewModel::class.java)
+            ViewModelProvider(this).get(ListViewModel::class.java)
 
         _binding = FragmentDashboardBinding.inflate(inflater, container, false)
         val root: View = binding.root
