+++ app/src/main/java/com/example/taskfrog/ui/list/ListViewModel.kt
@@ -1,10 +1,10 @@
-package com.example.taskfrog.ui.dashboard
+package com.example.taskfrog.ui.list
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class DashboardViewModel : ViewModel() {
+class ListViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is dashboard Fragment"
