+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -13,13 +13,13 @@
 
     <fragment
         android:id="@+id/navigation_dashboard"
-        android:name="com.example.taskfrog.ui.dashboard.DashboardFragment"
+        android:name="com.example.taskfrog.ui.list.ListFragment"
         android:label="@string/title_dashboard"
-        tools:layout="@layout/fragment_dashboard" />
+        tools:layout="@layout/fragment_list" />
 
     <fragment
         android:id="@+id/navigation_notifications"
-        android:name="com.example.taskfrog.ui.notifications.NotificationsFragment"
+        android:name="com.example.taskfrog.ui.notifications.CalendarFragment"
         android:label="@string/title_notifications"
-        tools:layout="@layout/fragment_notifications" />
+        tools:layout="@layout/fragment_calendar" />
 </navigation>
